requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
openpyxl==3.0.6
pycparser==2.20
pylightxl==1.52
PyNaCl==1.4.0
requests==2.25.1
selenium==3.141.0
six==1.15.0
soupsieve==2.1
undetected-chromedriver==2.1.2
urllib3==1.26.3
XlsxWriter==1.3.7
xlwt==1.3.0
